Add RSR and Ogrs calls to the community-api

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppstier/domain/ProtectScores.kt
@@ -4,13 +4,12 @@ import uk.gov.justice.digital.hmpps.hmppstier.domain.enums.AssessmentComplexityF
 import uk.gov.justice.digital.hmpps.hmppstier.domain.enums.ComplexityFactor
 import uk.gov.justice.digital.hmpps.hmppstier.domain.enums.Mappa
 import uk.gov.justice.digital.hmpps.hmppstier.domain.enums.Rosh
-import java.math.BigDecimal
 
 
 data class ProtectScores(
   val crn: String,
   val mappaLevel: Mappa?,
-  val rsrScore: BigDecimal?,
+  val rsrScore: Int?,
   val roshScore: Rosh?,
   val complexityFactors: List<ComplexityFactor>,
   val assessmentComplexityFactors: List<AssessmentComplexityFactor>
